resumeforge/requirements.txt
click>=8.1.7
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.2
google-generativeai>=0.3.0
rich>=13.7.0
